kg/pyproject.toml
[tool.poetry]
name = "kg"
version = "0.1.0"
description = ""
authors = ["awiradarma <awiradarma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
llama-index-llms-openai = "^0.1.20"
llama-index-embeddings-openai = "^0.1.10"
llama-index-graph-stores-nebula = "^0.1.3"
nebula3-python = "^3.8.0"
ipython = "^8.24.0"
llama-hub = "^0.0.79.post1"
wikipedia = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
